fix secret
